refactor(Name): 이름도 Name이라는 클래스에서 만들어지는 객체로 관리
객체지향적인 설계로 자동차 경주 게임이라는 도메인을 생각해봤을 때, 이름도 자동차에 속하게 되는
하나의 객체라고 판단, 클래스로 구성

InputView클래스에서 사용자의 이름을 입력받을 때 처리하던 이름에 대한 예외사항을 Name 클래스의
책임이라고 판단하여 위치를 옮김
