+++ README.md
@@ -29,7 +29,7 @@
 - 하나의 문자열로 이름들을 입력
     * [예외 처리] 입력이 없는 경우
 - 라운드 수를 입력
-    * [예외 처]리 입력이 없는 경
+    * [예외 처리] 입력이 없는 경우
     * [예외 처리] 숫자가 아닌 경우
     * [예외 처리] 0 이하의 숫자가 입력된 경우
 - 사용자가 입력한 이름 문자열을 (,) 기준으로 분리
